Add debug recipe for configuring inotifywait.
We have been experiencing unusual behavior which is resulting in the
tiered cluster tests to fail. This feels like a bug in the chef-server,
but we need more information about the events which are occurring
during installation and upgrade with the migration-level file.

This installs inotify-tools and creates a system service to help us
track the creation and modification of this file.